--- a/website/docs/r/compute_cluster.html.markdown
+++ b/website/docs/r/compute_cluster.html.markdown
@@ -465,10 +465,14 @@ details, see the referenced link in the above paragraph.
 * `vsan_dedup_enabled` - (Optional) Enables vSAN deduplication on the cluster.
   Cannot be independently set to true. When vSAN deduplication is enabled, vSAN
   compression is automatically enabled.
-* `vsan_compression_enabled` - (Optional) Enables vSAN compression on the cluster.
-* `vsan_performance_enabled` - (Optional) Enables vSAN performance service on the cluster. When vSAN is enabled, performance service will be enabled by default if it is not explicitly configured.
-* `vsan_verbose_mode_enabled` - (Optional) Enables verbose mode for vSAN performance service on the cluster.
-* `vsan_network_diagnostic_mode_enabled` - (Optional) Enables network diagnostic mode for vSAN performance service on the cluster.
+* `vsan_compression_enabled` - (Optional) Enables vSAN compression on the
+  cluster.
+* `vsan_performance_enabled` - (Optional) Enables vSAN performance service on
+  the cluster. Default: `true`.
+* `vsan_verbose_mode_enabled` - (Optional) Enables verbose mode for vSAN
+  performance service on the cluster.
+* `vsan_network_diagnostic_mode_enabled` - (Optional) Enables network
+  diagnostic mode for vSAN performance service on the cluster.
 * `vsan_disk_group` - (Optional) Represents the configuration of a host disk
   group in the cluster.
   * `cache` - The canonical name of the disk to use for vSAN cache.