@tenthirtyam 2022-08-25
After further analysis with the contributor, this race condition  is only observed in vSphere `.next`, but does not manifest in 7.0.x.  

This known issue shall be mitigated in a future update of the provider.

**Ryan Johnson**
Senior Staff Solutions Architect | Product Engineering @ VMware, Inc.